Make projection work properly
Don't convert projection argument to a string before adding it to @params,
because it will get converted again when the params are all converted to json.
Support (e.g.) projection(:gender) and projection(['gender', 'supplier_name'])
instead of assuming that if the projection argument is not a string,
it must be a hash.